requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
psycopg2==2.9.9
python-dotenv
Flask-SQLAlchemy
Flask-Migrate